create checkpoint dir only on rank 0

+++ src/fsdp/fsdp_hsdp_mnist.py
@@ -219,8 +219,10 @@ class Trainer:
             print_0(f"Test Average Loss: {loss:.4f}, Accuracy: {accuracy}")
 
     def save(self):
-        shutil.rmtree(self.checkpoint_dir, ignore_errors=True)
-        self.checkpoint_dir.mkdir(parents=True, exist_ok=True)
+        if dist.get_rank() == 0:
+            shutil.rmtree(self.checkpoint_dir, ignore_errors=True)
+            self.checkpoint_dir.mkdir(parents=True, exist_ok=True)
+
         dist.barrier()
         with FSDP.state_dict_type(self.model, StateDictType.SHARDED_STATE_DICT):
             model_state_dict = self.model.state_dict()
